lectures/part6/examples/requirements.txt
python-telegram-bot>=21.0
anthropic>=0.30.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
